requirements.txt
flask>=3.0
pillow>=10.0
numpy>=1.24
torch>=2.0
torchvision>=0.15
gunicorn>=21.0

# Optional — only needed for the TensorFlow/Keras model
tensorflow>=2.13
